# ws-java4web/ws-java4web-web/web-api/swagger/dist/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample project for workshop Java4Web\n\t\t\t\t\t\t\t\t\t(Foundation and Advanced).",
    "version" : "v1",
    "title" : "Java4Web Workshop Sample",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Mario Radan",
      "email" : "mario.radan@pentasys.de"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/api",
  "tags" : [ {
    "name" : "users"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/users/create/" : {
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Create user",
        "description" : "This can only be done by the logged in user.",
        "operationId" : "createUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "The user that needs to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ModelAndView"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Group" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "version" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "users" : {
          "type" : "array",
          "items" : {
            "type" : "object"
          }
        }
      }
    },
    "ModelAndView" : {
      "type" : "object",
      "properties" : {
        "view" : {
          "$ref" : "#/definitions/View"
        },
        "model" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "status" : {
          "type" : "string",
          "enum" : [ "CONTINUE", "SWITCHING_PROTOCOLS", "PROCESSING", "CHECKPOINT", "OK", "CREATED", "ACCEPTED", "NON_AUTHORITATIVE_INFORMATION", "NO_CONTENT", "RESET_CONTENT", "PARTIAL_CONTENT", "MULTI_STATUS", "ALREADY_REPORTED", "IM_USED", "MULTIPLE_CHOICES", "MOVED_PERMANENTLY", "FOUND", "MOVED_TEMPORARILY", "SEE_OTHER", "NOT_MODIFIED", "USE_PROXY", "TEMPORARY_REDIRECT", "RESUME_INCOMPLETE", "BAD_REQUEST", "UNAUTHORIZED", "PAYMENT_REQUIRED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "PROXY_AUTHENTICATION_REQUIRED", "REQUEST_TIMEOUT", "CONFLICT", "GONE", "LENGTH_REQUIRED", "PRECONDITION_FAILED", "REQUEST_ENTITY_TOO_LARGE", "REQUEST_URI_TOO_LONG", "UNSUPPORTED_MEDIA_TYPE", "REQUESTED_RANGE_NOT_SATISFIABLE", "EXPECTATION_FAILED", "I_AM_A_TEAPOT", "INSUFFICIENT_SPACE_ON_RESOURCE", "METHOD_FAILURE", "DESTINATION_LOCKED", "UNPROCESSABLE_ENTITY", "LOCKED", "FAILED_DEPENDENCY", "UPGRADE_REQUIRED", "PRECONDITION_REQUIRED", "TOO_MANY_REQUESTS", "REQUEST_HEADER_FIELDS_TOO_LARGE", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "GATEWAY_TIMEOUT", "HTTP_VERSION_NOT_SUPPORTED", "VARIANT_ALSO_NEGOTIATES", "INSUFFICIENT_STORAGE", "LOOP_DETECTED", "BANDWIDTH_LIMIT_EXCEEDED", "NOT_EXTENDED", "NETWORK_AUTHENTICATION_REQUIRED" ]
        },
        "empty" : {
          "type" : "boolean",
          "default" : false
        },
        "modelMap" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "reference" : {
          "type" : "boolean",
          "default" : false
        },
        "viewName" : {
          "type" : "string"
        }
      }
    },
    "User" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "version" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "firstname" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "active" : {
          "type" : "boolean",
          "default" : false
        },
        "group" : {
          "$ref" : "#/definitions/Group"
        }
      }
    },
    "View" : {
      "type" : "object",
      "properties" : {
        "contentType" : {
          "type" : "string"
        }
      }
    }
  }
}